added cartegories and more msmaging of products
