+++ backend/routes/admin.routes.ts
@@ -1,7 +1,8 @@
 import express from "express"
-import { createProduct } from "../controllers/createProduct.controller.js"
+import { createProduct, getAllProducts, getProductById } from "../controllers/Product.controller.js"
 
 const router = express.Router()
-
 router.post("/product", createProduct)
+router.get("/products", getAllProducts)
+router.get("/product/:id", getProductById)
 export default router
